# reborn_rentals_backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API – Proyecto X",
        "description": "Documentación de la API con autenticación JWT (Bearer).",
        "contact": {
            "email": "soporte@tudominio.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor base"
        }
    ],
    "paths": {
...
        "/api/paymentInfos": {
            "get": {
                "tags": [
                    "PaymentInfos"
                ],
                "summary": "Listado paginado (admin ve todas; usuario solo las suyas).",
                "operationId": "727529904ba8814880b3867718c90a4a",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "(Solo admin) filtra por usuario",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PaymentInfoResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/paymentInfo": {
            "post": {
                "tags": [
                    "PaymentInfos"
                ],
                "summary": "Crear una tarjeta (admin puede crear para cualquiera; user solo para sí mismo).",
                "operationId": "423d0e74cb44b057aa76a4714f1032b7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaymentInfoStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentInfoResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "422": {
                        "description": "Validación / duplicado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/paymentInfo/{paymentInfo}": {
            "get": {
                "tags": [
                    "PaymentInfos"
                ],
                "summary": "Ver una tarjeta por ID (admin cualquiera; user solo suya).",
                "operationId": "271a54865c263e78853e067c6b30cf9c",
                "parameters": [
                    {
                        "name": "paymentInfo",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentInfoResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "PaymentInfos"
                ],
                "summary": "Actualizar una tarjeta (admin cualquiera; user solo suya).",
                "operationId": "abfb585277d967e97328f2b4716a37be",
                "parameters": [
                    {
                        "name": "paymentInfo",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaymentInfoUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentInfoResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Validación / duplicado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "PaymentInfos"
                ],
                "summary": "Eliminar una tarjeta (admin cualquiera; user solo suya).",
                "operationId": "55e4473f9d7e94bac5fa27d432b5065a",
                "parameters": [
                    {
                        "name": "paymentInfo",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminado"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PaymentInfoResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 12
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 23
                    },
                    "card_holder_name": {
                        "type": "string",
                        "example": "Fernando Perez"
                    },
                    "card_number_mask": {
                        "type": "string",
                        "example": "**** **** **** 4242"
                    },
                    "card_expiration": {
                        "type": "string",
                        "example": "12/27"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "PaymentInfoStoreRequest": {
                "required": [
                    "user_id",
                    "card_holder_name",
                    "card_number",
                    "card_expiration"
                ],
                "properties": {
                    "user_id": {
                        "description": "Si no eres admin, debe ser tu mismo ID.",
                        "type": "integer",
                        "example": 23
                    },
                    "card_holder_name": {
                        "type": "string",
                        "example": "Fernando Perez"
                    },
                    "card_number": {
                        "type": "string",
                        "example": "4242424242424242"
                    },
                    "card_expiration": {
                        "type": "string",
                        "example": "12/27"
                    }
                },
                "type": "object"
            },
            "PaymentInfoUpdateRequest": {
                "properties": {
                    "user_id": {
                        "type": "integer",
                        "example": 23
                    },
                    "card_holder_name": {
                        "type": "string",
                        "example": "F. Perez"
                    },
                    "card_number": {
                        "type": "string",
                        "example": "4000000000000002"
                    },
                    "card_expiration": {
                        "type": "string",
                        "example": "01/28"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Envia el token como: Bearer {tu_token_jwt}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}